As far as I'm getting for today
I have most of the parse from Trello logic done, which in the
simplified form of TSync, that's most of the work (now it's just
checking for task completeness and then building the Todo.txt file.
I think I can get this up and running in a couple hours tomorrow.