Removed card/project tasks
The dependency thing in TaskWarrior was starting to get on my nerves.
Project tasks are considered more urgent than non-project tasks
regardless of how urgent they really are, and they somewhat dilute the
power of the TaskWarrior dependency system. So, I've started removing
the project task and replacing it with code to scan a card's tasks and
move it to "Done" if all of the tasks have been checked off.