Fixed choke on untagged TW tasks (Issue 6)
While building the TaskWarrior pending list, TSync was choking on tasks
that had no tags. This is because it was attempting to search through
tag list of a dictionary key that just wasn't there. Added some try/
except logic that caught cases where there were no tags, logs the find
as a warning, and then continues.